move getSRID and throw runtime exception if parsing fails
